requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
attrs==25.1.0
    # via
    #   cattrs
    #   requests-cache
babel==2.17.0
    # via flask-babel
blinker==1.9.0
    # via flask
cattrs==24.1.2
    # via requests-cache
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   click-default-group
    #   flask
    #   sqlite-utils
click-default-group==1.2.4
    # via sqlite-utils
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
dataclasses-json==0.6.7
    # via -r requirements.in
exceptiongroup==1.2.2
    # via
    #   cattrs
    #   pytest
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-babel
flask-babel==4.0.0
    # via -r requirements.in
graphqlclient==0.2.4
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
idna==3.10
    # via requests
iniconfig==2.0.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.5
    # via
    #   flask
    #   flask-babel
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
marshmallow==3.26.1
    # via dataclasses-json
mypy-extensions==1.0.0
    # via typing-inspect
narwhals==1.26.0
    # via plotly
packaging==24.2
    # via
    #   gunicorn
    #   marshmallow
    #   plotly
    #   pytest
platformdirs==4.3.6
    # via requests-cache
plotly==6.0.0
    # via -r requirements.in
pluggy==1.5.0
    # via
    #   pytest
    #   sqlite-utils
psycopg==3.2.4
    # via -r requirements.in
psycopg-binary==3.2.4
    # via psycopg
pytest==8.3.4
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via sqlite-utils
python-dotenv==1.0.1
    # via -r requirements.in
python-slugify==8.0.4
    # via -r requirements.in
pytz==2025.1
    # via flask-babel
requests==2.32.3
    # via
    #   -r requirements.in
    #   requests-cache
    #   requests-mock
requests-cache==1.2.1
    # via -r requirements.in
requests-mock==1.12.1
    # via -r requirements.in
ruff==0.9.6
    # via -r requirements.in
six==1.17.0
    # via
    #   graphqlclient
    #   python-dateutil
    #   url-normalize
sqlite-fts4==1.0.3
    # via sqlite-utils
sqlite-utils==3.38
    # via -r requirements.in
tabulate==0.9.0
    # via sqlite-utils
text-unidecode==1.3
    # via python-slugify
tomli==2.2.1
    # via pytest
tqdm==4.67.1
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   cattrs
    #   psycopg
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2025.1
    # via psycopg
url-normalize==1.4.3
    # via requests-cache
urllib3==2.3.0
    # via
    #   requests
    #   requests-cache
werkzeug==3.1.3
    # via flask
xlsxwriter==3.2.2
    # via -r requirements.in
